Added editing GUI's for the other components fixed a potential component dupe bug

+++ src/main/java/malte0811/industrialWires/blocks/controlpanel/TileEntityPanel.java
@@ -61,7 +61,7 @@ public class TileEntityPanel extends TileEntityIWBase implements IDirectionalTil
 	{
 		for (int i = 2; i < 14; i++) {
 			int color = EnumDyeColor.byMetadata(i - 2).getMapColor().colorValue;
-			IndicatorLight ind = new IndicatorLight(0, i - 2, color);
+			IndicatorLight ind = new IndicatorLight(0, (byte) (i - 2), color);
 			LightedButton btn = new LightedButton(color, false, true, 1, i - 2);
 			Label lbl = new Label("->", color);
 			ind.setX(0);
